# reports/discovered-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ESignPay API",
    "description": "API for managing document e-signatures, payments (Stripe 3D Secure), account opening with KYC/OCR verification, and OTP signing via Twilio SMS.\n\n### Compression\nAll endpoints support **gzip compression**. Send `Accept-Encoding: gzip` to receive compressed responses (typically 60\u201380% payload reduction). Responses \u2265 256 bytes are automatically compressed.\n\n### Range / Partial Content (206)\nFile download endpoints support **HTTP Range requests** (`Range: bytes=start-end`). The server responds with `206 Partial Content` and `Accept-Ranges: bytes`, enabling resumable downloads and partial fetches for large files.",
    "contact": {
      "name": "ESignPay Team"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "ESignPay API Server \u2014 gzip compression enabled on all endpoints",
      "x-server-compression": {
        "algorithms": [
          "gzip"
        ],
        "mime-types": [
          "application/json",
          "application/xml",
          "text/html",
          "text/plain",
          "application/hal+json",
          "application/cbor"
        ],
        "enabled": true,
        "min-response-size": "256 bytes"
      },
      "x-server-range-support": {
        "enabled": true,
        "unit": "bytes",
        "endpoints": [
          "/api/documents/{id}/download"
        ]
      },
      "x-server-etag-support": {
...
      }
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/sign/{token}": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Submit signature for a document (public)",
        "operationId": "signDocument",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDocumentResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/sign/{token}/verify-otp": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Verify OTP code (public)",
        "operationId": "verifyOtp",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOtpResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/sign/{token}/send-otp": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Send OTP code via SMS to signer (public)",
        "operationId": "sendOtp",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOtpResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List my documents (supports pagination & field filtering)",
        "operationId": "getMyDocuments",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-based)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            },
            "example": 20
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include (e.g. id,title,status)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageDocumentResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Create a new document",
        "operationId": "createDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/send": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Send document for signature",
        "operationId": "sendForSignature",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/resend": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Resend signature request to pending signers",
        "operationId": "resendForSignature",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sign/verify/{token}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Verify a signature token (public)",
        "operationId": "verifySignatureToken",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDocumentSignerResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        },
        "security": []
      }
    },
...
    "/api/documents/{id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get document detail (supports field filtering)",
        "operationId": "getDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDocumentResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete a document",
        "operationId": "deleteDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/documents/changes": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get documents changed since a given timestamp (delta sync)",
        "operationId": "getDocumentsChanges",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "ISO date-time (e.g. 2026-04-01T00:00:00)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "fields",
            "in": "query",
            "description": "Comma-separated list of fields to include",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Previous ETag value. If it matches, server returns 304.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "\"0a1b2c3d4e5f\""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "Content-Encoding": {
                "description": "gzip when client sends Accept-Encoding: gzip",
                "schema": {
                  "type": "string",
                  "enum": [
                    "gzip",
                    "identity"
                  ]
                }
              },
              "Vary": {
                "description": "Accept-Encoding",
                "schema": {
                  "type": "string",
                  "example": "Accept-Encoding"
                }
              },
              "ETag": {
                "description": "Entity tag for cache validation (MD5 of response body)",
                "schema": {
                  "type": "string",
                  "example": "\"0a1b2c3d4e5f\""
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListDocumentResponse"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 returned when If-None-Match matches current ETag. No body is sent, saving bandwidth."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SignDocumentRequest": {
        "type": "object",
        "properties": {
          "signatureData": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "signatureData"
        ]
      },
      "ApiResponseDocumentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentResponse"
          }
        }
      },
...
      "VerifyOtpRequest": {
        "type": "object",
        "properties": {
          "otpCode": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6
          }
        },
        "required": [
          "otpCode"
        ]
      },
...
      "SendOtpRequest": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\+[1-9]\\d{1,14}$"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Keycloak JWT Bearer token \u2014 obtain from POST to http://localhost:9090/realms/esignpayment/protocol/openid-connect/token",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}